Worked with adding and deleting photos. Fixed button bugs.

+++ src/app/components/my-account/bikes-info/bike/bike.component.ts
@@ -1,8 +1,16 @@
-import { Component, OnInit, Input } from '@angular/core';
+import { Component, OnInit, Input, Output, EventEmitter} from '@angular/core';
 import { brakes, brands, frames, categories, sizes, diameterWheels } from './constantsBikeInfo';
 import { BikeInfo } from '../../../../models/bike-info.model';
 import { BikeInfoService } from '../../../../services/bike-info.service';
 import { FormGroup, FormBuilder, Validators } from '@angular/forms';
+import { MatSnackBar } from '@angular/material/snack-bar';
+
+class ImageSnippet {
+  pending: boolean = false;
+  status: string = 'init';
+
+  constructor(public src: string, public file: File) {}
+}
 
 @Component({
   selector: 'app-bike',
@@ -11,10 +19,12 @@ import { FormGroup, FormBuilder, Validators } from '@angular/forms';
 })
 export class BikeComponent implements OnInit {
   @Input() bike: BikeInfo;
+  @Output() removeLastEmptyBike = new EventEmitter<boolean>();
 
   private id: string = '';
   submitted: boolean = false;
   constants = {brakes, brands, frames, categories, sizes, diameterWheels};
+  images: string[] = [];
 
   bikeForm: FormGroup = this.formBuilder.group({
     brand: ['', Validators.required],
@@ -30,7 +40,13 @@ export class BikeComponent implements OnInit {
     price_rent: ['', Validators.required],
   });
 
-  constructor(private formBuilder: FormBuilder, private bikeInfoService: BikeInfoService) {}
+  selectedFile: ImageSnippet;
+
+  constructor(
+    private formBuilder: FormBuilder,
+    private bikeInfoService: BikeInfoService,
+    private snackBar: MatSnackBar
+  ) {}
 
   ngOnInit() {
     if (this.bike) {
@@ -49,6 +65,7 @@ export class BikeComponent implements OnInit {
       });
 
       this.id = this.bike.id || '';
+      this.images = this.bike.images || [];
     }
   }
 
@@ -63,13 +80,78 @@ export class BikeComponent implements OnInit {
     }
 
     if (this.id) {
-      this.bikeInfoService.updateBikeItem(this.id, this.bikeForm.value);
+      this.bikeInfoService.updateBikeItem(this.id, this.bikeForm.value)
+        .then((res) => {
+          this.onSuccess(`Bike Info saved succesfuly!`);
+        },
+        (err) => {
+          this.onError(`Bike Info Failed!`);
+        });
     }
 
     this.bikeInfoService.pushBikeItem(this.bikeForm.value);
   }
 
   delete() {
-      this.bikeInfoService.deleteBikeItem(this.id);
+    if (this.id) {
+      this.bikeInfoService.deleteBikeItem(this.id)
+        .then((res) => {
+          this.onSuccess(`Bike removed succesfuly!`);
+        },
+        (err) => {
+          this.onError(`Bike not removed. Failed!`);
+        });
+    } else {
+      this.removeLastEmptyBike.emit(true);
+    }
+  }
+
+  removeImage(removeImage) {
+    this.bikeInfoService.updateBikeItem(this.id, { images: this.images.filter((image: string) => image !== removeImage) })
+      .then((res) => {
+        this.onSuccess(`Image Removed Succesfuly!`);
+      },
+      (err) => {
+        this.onError(`Image Removed Failed!`);
+      })
+  }
+
+  processFile(imageInput: any) {
+    const file: File = imageInput.files[0];
+    const reader = new FileReader();
+
+    reader.addEventListener('load', (event: any) => {
+
+      this.selectedFile = new ImageSnippet(event.target.result, file);
+      this.selectedFile.pending = true;
+
+      this.images.push(this.selectedFile.src)
+
+      this.bikeInfoService.updateBikeItem(this.id, { images: this.images })
+        .then((res) => {
+            this.onSuccess(`Image Uploaded Succesfuly!`);
+          },
+          (err) => {
+            this.onError(`Image Upload Failed!`);
+          })
+    });
+
+    reader.readAsDataURL(file);
+  }
+
+  private onSuccess(message) {
+    this.snackBar.open(message, '', {
+      duration: 2000,
+      panelClass: ['succesful-snackbar'],
+      horizontalPosition: "right"
+    }); 
+  }
+
+  private onError(message) {
+    this.snackBar.open(message, '', {
+      duration: 2000,
+      panelClass: ['error-snackbar'],
+      horizontalPosition: "right"
+    }); 
   }
 }
